Backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
psycopg==3.2.12
psycopg-binary==3.2.12
sqlparse==0.5.3
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
